Refactor: Product 생성 로직 분리
