+++ src/main/java/org/example/sivillage/review/application/ReviewService.java
@@ -4,8 +4,8 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.example.sivillage.BeautyInfo.domain.BeautyInfo;
 import org.example.sivillage.BeautyInfo.infrastructure.BeautyInfoRepository;
+import org.example.sivillage.product.application.ProductServiceImpl;
 import org.example.sivillage.review.domain.CategoryType;
-import org.example.sivillage.product.application.ProductService;
 import org.example.sivillage.review.domain.Review;
 import org.example.sivillage.review.domain.ReviewImage;
 import org.example.sivillage.review.dto.ReviewRequestDto;
@@ -27,7 +27,7 @@ public class ReviewService {
     private final ReviewImageRepository reviewImageRepository;
     private final BeautyInfoRepository beautyInfoRepository;
     private final SizeInfoRepository sizeInfoRepository;
-    private final ProductService productService;
+    private final ProductServiceImpl productServiceImpl;
 
     public void addReview(ReviewRequestDto dto,  String authorEmail, String memberUuid, String productUuid) {
 
@@ -35,7 +35,7 @@ public class ReviewService {
         SizeInfo sizeInfo = sizeInfoRepository.findByMemberUuid(memberUuid).orElse(new SizeInfo());
         Review review = Review.toEntity(dto, authorEmail, memberUuid, productUuid);
 
-        String info = CategoryType.fromCategoryPath(productService.getCategoryPath(productUuid).getCategoryPath())
+        String info = CategoryType.fromCategoryPath(productServiceImpl.getCategoryPath(productUuid).getCategoryPath())
                         .getInfo(beautyInfo, sizeInfo);
 
         review.toEntityMemberInfo(info);
